minor approval voting refactor
